Extract `TypeSelector` component
